Added persisted flags to SongEntity, AlbumEntity, and ArtistEntity.
I'm still mulling it over if this is the best way to store the persisted
state of each of these, but this is where I'm going with it for now.
This still needs calls in SongDao and AlbumDao to work out transitive
persistence (i.e. if a song's album or artist are persisted, or if an
album's artist is persisted).