Overhauled BaseRequest a bit.
BaseRequest now has utility methods to extract the basic OpenSubsonic
data from a JSON response and to throw an OpenSubsonicException if any
sort of error occurred.

Also, execute() is no longer part of BaseRequest for the time being.  As
it stands, each request will be able to return a variety of data (i.e.
ping will just return the basic response data; search3 needs to return
arrays of artists, albums, and songs; any binary fetcher will return
either the binary data or a filehandle OR basic response data if there
was an error), so trying to make them share a single method signature is
fraught with peril.  I may change my mind later if I can think of a
clean way around it.