Changed BaseRequestData to take a Server, not individual components.
I realized that any request that will get data to be stored for later
(such as fetching the song database) will need to have the server ID to
stuff it in the database.  Rather than add more params to
BaseRequestData, it now just takes a Server object.

Similarly, there is now a static companion function to make a LoginData
from a Server, and makeBaseUriBuilder() calls that.