Overrode toString on Server.
Since a Server object contains password and/or API key data, it probably
shouldn't be doing its default toString, which might wind up dumping
said data in plaintext in the logs.  This overrides it to hide that
data.

With that, LoginData.makeLoginDataForServer can now just output the
Server object itself if something goes weird.